added js repl, added exe subcommand to pkg, bugfixes

+++ src/js/sys/builtin.js
@@ -1,7 +1,7 @@
 import { directoryAutoComplete, enumAutoCompleteFactory, fileAutoComplete } from "./autocomplete.js";
 import { EXECUTION_PATH_FOLDER } from "./const.js";
 import { fs, rootfs } from "./fs.js"
-import { runScript } from "./run_script.js";
+import { runCode, runScript } from "./run_script.js";
 
 export const terminalBuiltin = {
     /**
@@ -52,6 +52,21 @@ export const terminalBuiltin = {
 	},
 	async df() { await this.storage() },
 	async js(path) {
+        // js repl
+        if(path === undefined) {
+            this.terminal.log("JS REPL\nAuto included packages: fs, console\ntype 'exit' to exit\n")
+            while(true) {
+                const input = await this.terminal.prompt('>>>');
+                if(input === 'exit') return;
+
+                // include important packages like console and fs automatically 
+                // so this repl is somewhat useful
+                const INCLUDE_CODE = 'const console=include("console"),fs=include("fs");';
+                const result = await runCode(this.terminal, INCLUDE_CODE + input);
+                if(result !== undefined) this.terminal.log(result);
+            }
+        }
+
 		const resolved = fs.resolvePath(path, this.terminal.path);
 		runScript(this.terminal, resolved);
 	},
@@ -64,16 +79,16 @@ export const terminalBuiltin = {
 	 * @param {string} script 
 	 * @returns 
 	 */
-	async pkg(subcommand, script) {
-		if(!pkgManager.hasOwnProperty(subcommand)) throw new Error(`package manager does not have subcommand '${subcommand}'\nValid subcommands: ${extractCommands(pkgManager)}`);
-        pkgManager[subcommand](script);
-	},
+        async pkg(subcommand, script) {
+            if(!pkgManager.hasOwnProperty(subcommand)) throw new Error(`package manager does not have subcommand '${subcommand}'\nValid subcommands: ${extractCommands(pkgManager).join(' ')}`);
+            pkgManager[subcommand].bind(this)(script);
+        },
     async help() {
         // show list of commands 
         this.terminal.log('list of commands:');
         this.terminal.log(extractCommands(this).join(' '));
         
-        const dir = await fs.getDirectory('/src');
+        const dir = await fs.getDirectory('/src/cmd');
         const arr = [];
         for await (const [key, value] of dir.entries()) {
             if(key.slice(key.length - 3) === '.js') {
@@ -83,7 +98,8 @@ export const terminalBuiltin = {
 
         this.terminal.log(arr.join(' '));
     },
-    async rm(path) {
+    async rm(path = null) {
+        if(path === null) throw new Error('expected 1 argument, got 0')
         await fs.deleteFile(path, this.terminal.path);
         this.terminal.log('successfully deleted ' + fs.stringifyPath(fs.resolvePath(path, this.terminal.path)))
     }
@@ -105,6 +121,17 @@ const pkgManager = {
     async remove(script) {
         if(script === undefined || !script.match(/^[A-Za-z0-9_+-\.]+$/)) return this.terminal.log('Invalid script name');
         fs.deleteFile(`${EXECUTION_PATH_FOLDER}${script}.js`);
+    },
+    async exe(script) {
+        if(script === undefined || !script.match(/^[A-Za-z0-9_+-\.]+$/)) return this.terminal.log('Invalid script name');
+		const path = `/src/js/cmd/${script}.js`;
+		this.terminal.log('GET ' + location.origin + path);
+        
+        const response = await fetch(path);
+        if(response.status === 404) return this.terminal.log('file not found')
+        const text = await response.text();
+        this.terminal.log('executing downloaded file')
+        runCode(this.terminal, text);
     }
 }
